tech(api): migrate delete-unreferenced-translations script
